requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgiref==3.3.1
aspy.yaml==1.3.0
astroid==2.4.2
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
backports.csv==1.0.7
bandit==1.6.2
black==19.10b0
certifi==2019.6.16
cfgv==2.0.1
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.4
confusable-homoglyphs==3.2.0
coverage==4.5.3
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap3==14.2.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-registration==3.1.1
django-tables2==2.3.3
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.8.4
gitdb==4.0.5
GitPython==3.1.11
gunicorn==19.9.0
identify==1.4.7
idna==2.8
importlib-metadata==1.7.0
importlib-resources==1.0.2
ipython==7.6.0
ipython-genutils==0.2.0
isort==5.6.3
jdcal==1.4.1
jedi==0.14.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-mommy==1.6.0
more-itertools==7.1.0
nodeenv==1.3.3
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
parso==0.5.0
pathspec==0.8.1
pbr==5.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
pre-commit==1.18.3
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.4.2
pylint==2.6.0
pyparsing==2.4.0
pytest==5.0.0
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
regex==2020.11.13
requests==2.22.0
six==1.12.0
smmap==3.0.4
sqlparse==0.3.0
stevedore==3.3.0
tablib==0.13.0
toml==0.10.0
tqdm==4.32.2
traitlets==4.3.2
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
virtualenv==16.7.5
wcwidth==0.1.7
whitenoise==4.1.3
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.1
